+++ app/src/main/java/com/stephenwanjala/multiply/game/components/AnimatedBackground.kt
@@ -46,7 +46,6 @@ fun AnimatedBackground() {
                         translationX = position.x
                         translationY = position.y
                     }
-//                    .offset { IntOffset(position.x.toInt(), position.y.toInt()) }
             )
 
             LaunchedEffect(Unit) {
