+++ app/src/main/java/com/stephenwanjala/multiply/game/models/BubbleMathDifficulty.kt
@@ -1,0 +1,7 @@
+package com.stephenwanjala.multiply.game.models
+
+enum class BubbleMathDifficulty {
+    EASY,
+    MEDIUM,
+    HARD
+}
