+++ app/src/main/java/com/stephenwanjala/multiply/game/models/GameMode.kt
@@ -1,0 +1,6 @@
+package com.stephenwanjala.multiply.game.models
+
+sealed class GameMode {
+    data class BubbleMathBlitz(val difficulty: BubbleMathDifficulty) : GameMode()
+    data class QuizGenius(val difficulty: QuizDifficulty) : GameMode()
+}
